+++ decnumber-sys/src/lib.rs
@@ -1092,7 +1092,7 @@ extern "C" {
     pub fn decimal128Canonical(arg1: *mut decimal128, arg2: *const decimal128) -> *mut decimal128;
 }
 
-/// Constants used for efficiently decoding decimal coefficients.
+// Constants used for efficiently decoding decimal coefficients.
 extern "C" {
     pub static DPD2BIN: [u16; 1024];
     pub static DPD2BINK: [u32; 1024];